Blacklist from menus.
Summary: {D8} was an attempt to do the same thing by creating a whole menu with an explicit blacklist. While this should have worked, it didn't. While it doesn't seem to have adverse effects, it could. Therefore, a more functional and simple solution and one that other non-Lubuntu users could benefit from is not showing it ever. Here we edit the .desktop file to get that resolved.

Test Plan: Make sure nm-tray is not in the menu.

Reviewers: tsimonq2

Reviewed By: tsimonq2

Differential Revision: https://phab.lubuntu.me/D14